+++ react/features/base/ui/functions.web.js
@@ -48,22 +48,3 @@ export function formatCommonClasses(stylesObj: Object) {
     return formatted;
 }
 
-/**
- * Overwrites recursively values from object 2 into object 1 based on common keys.
- * (Merges object2 into object1).
- *
- * @param {Object} obj1 - The object holding the merged values.
- * @param {Object} obj2 - The object to compare to and take values from.
- * @returns {void}
- */
-export function overwriteRecurrsive(obj1: Object, obj2: Object) {
-    Object.keys(obj2).forEach(key => {
-        if (obj1.hasOwnProperty(key)) {
-            if (typeof obj1[key] === 'object') {
-                overwriteRecurrsive(obj1[key], obj2[key]);
-            } else {
-                obj1[key] = obj2[key];
-            }
-        }
-    });
-}
