+++ react/features/dynamic-branding/middleware.js
@@ -3,7 +3,9 @@
 import { APP_WILL_MOUNT } from '../base/app';
 import { MiddlewareRegistry } from '../base/redux';
 
+import { SET_DYNAMIC_BRANDING_DATA } from './actionTypes';
 import { fetchCustomBrandingData } from './actions';
+import { createMuiBrandingTheme } from './functions';
 
 MiddlewareRegistry.register(store => next => action => {
     switch (action.type) {
@@ -11,6 +13,13 @@ MiddlewareRegistry.register(store => next => action => {
 
         store.dispatch(fetchCustomBrandingData());
         break;
+    }
+    case SET_DYNAMIC_BRANDING_DATA: {
+        const { customTheme } = action.value;
+
+        if (customTheme) {
+            action.value.muiBrandedTheme = createMuiBrandingTheme(customTheme);
+        }
     }
     }
