+++ react/features/dynamic-branding/reducer.js
@@ -101,6 +101,14 @@ const DEFAULT_STATE = {
     logoImageUrl: '',
 
     /**
+     * The generated MUI branded theme based on the custom theme json.
+     *
+     * @public
+     * @type {boolean}
+     */
+    muiBrandedTheme: undefined,
+
+    /**
      * The lobby/prejoin background.
      *
      * @public
@@ -135,12 +143,12 @@ ReducerRegistry.register(STORE_NAME, (state = DEFAULT_STATE, action) => {
             avatarBackgrounds,
             backgroundColor,
             backgroundImageUrl,
-            customTheme,
             defaultBranding,
             didPageUrl,
             inviteDomain,
             logoClickUrl,
             logoImageUrl,
+            muiBrandedTheme,
             premeetingBackground,
             virtualBackgrounds
         } = action.value;
@@ -149,12 +157,12 @@ ReducerRegistry.register(STORE_NAME, (state = DEFAULT_STATE, action) => {
             avatarBackgrounds,
             backgroundColor,
             backgroundImageUrl,
-            customTheme,
             defaultBranding,
             didPageUrl,
             inviteDomain,
             logoClickUrl,
             logoImageUrl,
+            muiBrandedTheme,
             premeetingBackground,
             customizationFailed: false,
             customizationReady: true,
